webhook: refactor to use controller-runtime server (#1751)
Refactor the webhook to use the controller-runtime provided
webhook.Server instead of a plain HTTP server. The biggest benefit this
brings, is it automatically loading changed certificate files instead of
forever using the ones present when the server was started.

This could have been implemented by using the controller-runtime
CertWatcher manually, but that would've been more code and aligning
machine-controller with the ecosystem-provided standards seems worth it.

Signed-off-by: Mara Sophie Grosch <mgrosch@anexia-it.com>